% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_enrich2.r
\name{r2_enrich}
\alias{r2_enrich}
\title{r2_enrich}
\usage{
r2_enrich(dat, v1, v2, nv, exp1)
}
\arguments{
\item{dat}{N by (M+1) matrix having variables in the order of cbind(y,x)}

\item{nv}{sample size}

\item{exp1}{The expectation of the ratio (e.g. # SNPs for the genomic region of interest / total # SNPs in genomic partitioning)}

\item{v1/v2}{These can be set as v1=1 and v2=2, v1=2 and v2=1, v1=3 and v2=2, or any combination as long as the value is between 1 - M}
}
\description{
This function estimates var(t1/(t1+t2))
where t1 = R2(y~x[,v1]+x[,v2]) - R2(y~x[,v1])) and
t2 = R2(y~x[,v1]+x[,v2]) - R2(y~x[,v2]))
where R2 is the R squared value of the model,
y is N by 1 matrix having the dependent variable, and
x is N by M matrix having M explanatory variables.
v1 or v2 indicates the ith column in the x matrix
(v1 or v2 should be a single interger between 1 - M, see Arguments below)
}
\keyword{R2}
\keyword{between}
\keyword{of}
\keyword{ratio}
\keyword{variance}
